requirements.txt
numpy
scipy
pillow
dahuffman
matplotlib
opencv-python

